vista info-profesor-view implementada

+++ src/main/java/com/cesur/aplicaciondual/domain/entities/profesor/ProfesorDAO.java
@@ -4,7 +4,7 @@ import java.util.ArrayList;
 
 public interface ProfesorDAO {
     public ArrayList<Profesor> getAll();
-    public Profesor get(Long id);
+    public Profesor get(Integer id);
     public Profesor save(Profesor data);
     public void update(Profesor data);
     public void delete(Profesor data);
